move mapper consent component to its own file
